Upload function creates SeFile instance and saves to database, or updates instance updated_at time if it already exists.

+++ se_filesystem/app/controllers/files_controller.rb
@@ -10,10 +10,41 @@ class FilesController < ApplicationController
   end
 
   def upload
-    #@file = SeFile.new(name: params[:name])
-    File.open("../" + params[:name],"w") do |f|
-      f.write(params[:data])
+
+    #Need to make sure it will not overwrite if user does not have access
+
+
+    #If file already exists in database, update it
+    if SeFile.exists?(name: params[:name])
+
+      @file = SeFile.find_by(name: params[:name])
+      @file.update_attribute(:updated_at, DateTime.now)
+
+      #write to file
+      File.open("../" + params[:name],"w") do |f|
+        f.write(params[:data])
+      end
+
+      render :text => "File Updated"
+
+    else #If file doesn't exist in database, create it
+
+      @file = SeFile.new(name: params[:name])
+
+      if @file.save
+
+        #write to file
+        File.open("../" + params[:name],"w") do |f|
+          f.write(params[:data])
+        end
+
+        render :text => "File Created"
+
+      else
+
+        render :text => "File not uploaded."
+
+      end
     end
-     render :text => "Success"
   end
 end
